# webapp/.well-known/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Root", "description": "Allows to open the API documentation in the browser directly instead of\nrequiring to open the /docs path.", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/countries": {"get": {"summary": "Countries", "description": "Returns a list of available countries.", "operationId": "countries_countries_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/countries/{country}": {"get": {"summary": "Cities", "description": "Returns a list of cities for a given country.", "operationId": "cities_countries__country__get", "parameters": [{"name": "country", "in": "path", "required": true, "schema": {"type": "string", "title": "Country"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/countries/{country}/{city}": {"get": {"summary": "Monthly Averages", "description": "Returns the monthly averages for a given city in a country.", "operationId": "monthly_averages_countries__country___city__get", "parameters": [{"name": "country", "in": "path", "required": true, "schema": {"type": "string", "title": "Country"}}, {"name": "city", "in": "path", "required": true, "schema": {"type": "string", "title": "City"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/countries/{country}/{city}/{month}": {"get": {"summary": "Monthly Average", "description": "Returns the average temperature for a given city in a country for a specific month.", "operationId": "monthly_average_countries__country___city___month__get", "parameters": [{"name": "country", "in": "path", "required": true, "schema": {"type": "string", "title": "Country"}}, {"name": "city", "in": "path", "required": true, "schema": {"type": "string", "title": "City"}}, {"name": "month", "in": "path", "required": true, "schema": {"type": "string", "title": "Month"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}